Make the sparse POA actually sparse.
This entailed:

  - changing the Interval type to pair<int,int> (signed) to
    accommodate the negative value used in empty interval
    endpoint;

  - fixing broken signed/unsigned comparisons;

  - respecting the beginRow, endRow bounds passed in from the
    RangeFinder;

  - properly handling the edge case of "no anchors found".

Additionally, useful diagnostic output has been added, hidden behind
the DEBUG_RANGE_FINDER flag, in case we need to troubleshoot a bug in
the future.

+++ include/pacbio/consensus/poa/RangeFinder.h
@@ -35,7 +35,7 @@ class PoaGraphImpl;
 class SdpRangeFinder
 {
 private:
-    std::map<PoaGraph::Vertex, std::pair<size_t, size_t>> alignableReadIntervalByVertex_;
+    std::map<PoaGraph::Vertex, std::pair<int, int>> alignableReadIntervalByVertex_;
 
 public:
     virtual ~SdpRangeFinder();
@@ -44,7 +44,7 @@ public:
                          const std::vector<PoaGraph::Vertex>& consensusPath,
                          const std::string& consensusSequence, const std::string& readSequence);
 
-    std::pair<size_t, size_t> FindAlignableRange(PoaGraph::Vertex v);
+    std::pair<int, int> FindAlignableRange(PoaGraph::Vertex v);
 
 protected:
     virtual SdpAnchorVector FindAnchors(const std::string& consensusSequence,
